rename the "modules" directory to "pages" to avoid confusion with terms
