+++ codegraph-viewer/frontend/app/pages/artifacts/artifactDetails.tsx
