+++ codegraph-viewer/frontend/app/pages/artifacts/artifacts.tsx
