+++ codegraph-viewer/frontend/app/pages/viewer/displayOptions.tsx
